+++ src/page/mod.rs
@@ -1,3 +1,5 @@
+mod main;
 mod tournament;
 
+pub use main::Main;
 pub use tournament::Tournament;
