+++ src/route.rs
@@ -1,0 +1,7 @@
+use yew_router::prelude::*;
+
+#[derive(Clone, Routable, PartialEq, Copy)]
+pub enum Route {
+  #[at("/")]
+  MainPage,
+}
